requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.9.0
celery==5.6.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
Django==6.0.4
django-admin-interface==0.32.0
django-ckeditor-5==0.2.14
django-colorfield==0.14.0
django-environ==0.13.0
django-js-asset==3.1.2
django-solo==2.5.1
django-unfold==0.89.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
redis==7.4.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ujson==5.12.0
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.3
google-auth==2.38.0
requests==2.32.3
